api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
azure-functions
PyJWT==2.9.0
qrcode==7.4.2
requests==2.32.3
